no more int Mode, replaced with Utils.Transformations mode

+++ Assets/Scripts/HandleGUI.cs
@@ -24,20 +24,20 @@ public class HandleGUI : MonoBehaviour {
         selectTranslate.SetActive(true);
         selectRotate.SetActive(false);
         selectScale.SetActive(false);
-        MainController.control.transformationNow = (int)Utils.Transformations.Translation;
+        MainController.control.transformationNow = Utils.Transformations.Translation;
     }
     public void buttonRotate() {
         selectTranslate.SetActive(false);
         selectRotate.SetActive(true);
         selectScale.SetActive(false);
-        MainController.control.transformationNow = (int)Utils.Transformations.Rotation;
+        MainController.control.transformationNow = Utils.Transformations.Rotation;
 
     }
     public void buttonScale() {
         selectTranslate.SetActive(false);
         selectRotate.SetActive(false);
         selectScale.SetActive(true);
-        MainController.control.transformationNow = (int)Utils.Transformations.Scale;
+        MainController.control.transformationNow = Utils.Transformations.Scale;
 
     }
